[BGs] Averages appear done
Three methodologies are included, none of which are tested.

Simple average: Each entry is equally important.
Time-weighted average: Each entry is weighed based on how much time is
between it and other entries.
"Camera flash" average: Each entry is equally important, except for
those within 10 minutes of preceding entry, which are logarithmically
less important. (Thanks plugnplay for the math.)

The "camera flash" average is selected in this commit.